Update 8/26
Spent majority of time trying to find work around so that require() could use variable path, did not work.

Need to find a way for the sidebar to effect the state of the carousel, probably use <Outlet /> in place of {children} so that Main can pass props. Would like to find a way to temporarily decrease interval instead or just select the appropriate button.