8/27 Update
Split content into Sidebar and CarouselContent so that props can be passed up between the two.

Added ModelImage component to pop up when a poster is clicked. Need to figure out how to:
A) Pause the Carousel when a modal is shown, most likely have to revert to regular bootstrap instead of react-bootstrap.
B) Figure out why the modal is showing all the images in reverse order, instead of just the one for the clicked image, probably have to use index and state as modals may run outside of normal operations?
(C)Add alt text and other attributes to ModalImage.

Figure out how to use the URL as a resource for carousel by tracking what image it's currently on without rendering.