9/1
Attempted to adjust carousel so that the image doesn't extend past view height or width. I got to work then messed with some stuff and now its a mess. Will most likely have to go back through the css file and start over as well as double check the containers and props for the carousel.

Need to push whenever I get something working in the future.