Moved getImages to Layout.js
Because the sidebar is going to be in different components within Layout.js depending on viewport (navbar for mobile) and it's going to be used to navigate to dynamic assets depending on path, in order for props to be managed correctly assets must be fetched within the Layout component. Changes were made accordingly.

Additionally, some steps were taken to start an offcanvas rendering of the sidebar for mobile.