Carousel Sidebar working
Got the sidebar to go to the appropriate slide when clicked. Still need to change link container size to fit to the view height, might be easier just to select overflow as hidden.

Changed how the asset prop is passed down to children.

Moved carousel index state to Layout so that when the sidebar is within Nav  for mobile it can still be tracked.

Moved sidebar for nonmobile to Content so that the content and sidebar are within the same row.