JSON server running, quality of life changes
JSON server is up and running.

changed path for Calendar component database.

Added Outlet possibility to Calendar to change how it is rendered

Utilized URL path for which calendar year and month are rendered.

Added getCurrentDate and formatMonth functions to API as well as implemented them within the app.

Fixed loadCalendarDb, could use some tuning, will need to move the majority of the function to api and streamline what is returned. Need to add month to the fetch path so that it only grabs whats needed.