Changed extensions and started building calender
Due to errors occurring to resetCalendar due to importing formatMonth, it had to be structured to .mjs extension along with api.js. This then required the calendar_db.json to have to be moved into the src folder.

Began constructing the calendar, however the columns are not aligning to the appropriate day of the week and are all within the Sunday col. Still needs a lot of work, but out of time. Will continue to progress.