Refined variable names and Nav.js, fixed Books.js error
Due to Books being an outlet for Layout it wouldn't rerender the outlet after loading assets for it, in this case the images for the carousel. Tried using useEffect, didn't work. So it was fixed in Nav.js by setting reloadDocument to true and making the assets a const variable instead.

Altered Nav to utilize a map function to list out the Nav links in place of manually entering. Utilizes an array of objects for necessary info for each individual link.