Adjusted api, Layout, and how sidebars function
Mainly adjusted for read ability and DRY code. Allowed Main to include sidebar instead of referencing it on each individual page.

This adjustment required changes to how Sidebar functioned where it is used to select what type of sidebar is used, instead of containing the code itself.

Gave each page a sidebar object in loadMultipleFiles to automatically select sidebar data points and the type of sidebar used.