General completion of Games page and simplification of code
Rather than make every page of the website mostly dependent on itself for code, adjusted components so that each page needs to only pass a few variables and whatever add-ons unique to the page to render it.

To be more specific Sidebar was added to Main and the type and source of sidebar content is decided at loadMultipleFiles along with the format of each page' s content. format folder was added, containing the various general components for each page's content.

General components for each page were created/refined and use switch cases to decide what type of elements will be rendered, along with their content. Additionally, using strict file names for assets and how loadMultipleFiles operates allowed for sharing object keys for rendering disparate content using a single component.

To a lesser extent carousel's were adjusted for a cleaner look.

Games was generally completed with more refinement necessary and Books and Music were adjusted to account for changes.