Partial completion of styling, dry fixes, bug with ContactMe
General update on style of web pages, incomplete.

ContactMe.js is not rendering the buttons as they should, justifyContent seems to be a suggestion. Still working on fix. Added Links.js to help with readability of ContactMe.js

Changed some descriptions for text elements within the pages.

Made Sidebar more dry.

Deleted some of the unnecessary styles in App.css, need to rework still to format for easier read.

Home.js had an additional header element, replaced with h3.

Added a button to Book.js to change format from carousel to horizontal list, as a horizontal list looks better, however it only required having assets as a state in Layout.js and adding a function to change it to change back and forth, so that was done instead.

Tooltips were added to Nav.js