added banner, minor changes to content
Banner for the header still isn't complete, but it was set up for when a good image is made/found.

Adjusted the content for MoviesTV and Games. Added Alice in Wonderland for movies. Games added hyperlinks to some of the content related to one of them.

Adjusted Item.js => cards. Added the ability to include hyperlinks at the end of the description. Attempted to be able to replace "<links>" string within the description with hyperlinks, but it broke the elements, rendering [object, Object] in place of hyperlink and I'm unsure how to designate it as html script.

Made some style and layout changes to TMDB acknowledgement.

Added to assets and adjusted loadMultipleFiles to create Header assets.