Misc. fixes for readability/DRY, package update, misc. bug fixes, mobile update
Instead of using useOutletContext() to pass props deeply nested, changed to useContext(). This enabled the ability to pass props to the sidebar as well as the main content. Removed unnecessary props, instances of useOutletContext() and replaced with useContext for relative components. Created a component, src/layout/PropContext.js to store context data.

Added header asset file to more easily implement change in the banner and other assets once those are ready.

Changed some of the standard html elements to components from frameworks such as <a> to <Link> from react-router-dom.

src\layouot\MobileToolbar.js is generally completed.

src\layout\Nav.js has been updated with mobile in mind as well as styling. Having problems with getting the navbar to stick like with the sidebar on desktop, I'm unsure what is causing the bug