Footer completion,  style changes, general clean up
src/layout/Footer.js was completed

App.css was organized into groups.

in-line styling was removed, unless only one key, value pair was used, and replaced with a className.

classNames were formatted to kebab case and some were renamed to be more clear.

added emailFunction() to api so that it can be accessed by both the Footer.js and ContactMe.js components.

Banner was updated.

Links was moved to src/layout because it is accessed by Footer.js.