Update for GH pages
Due to what I'm assuming is the hashrouter operating differently than browserrouter, not reloading the document due to changing hash on navigation despite setting to true,  the document no longer updates window.location on navigation. Additionally, window.location.hash is required instead of pathname to get relevant data to load content.

As such code was updated to use state for the URL data and adjustments were made to get the hash data as well as update its state as necessary.