Bug fix, infinite re-render or not loading content unless page was reloaded
It is believed the bug came about due to two simultaneous state updates, which caused the infinite re-render when code functioned in spite of this. When code was out of sync due to the state updates it resulted in a blank screen.

Bug was fixed by removing one of the states and replacing it with a constant variable.

Due to using hashrouter, using href=`#${id}` to scroll to id no longer works, replaced with appropriate function to scroll to id.