Bug fix, no state update on manual URL change, Error page update
Added origin.json file to all relevant folders in src/assets. This allowed a function to check if the hash location matched the loaded assets, if not reloaded the page to update state. This fixed a bug that would prevent state updates if the user manually entering a web address after initial entering the site due to using a hashrouter.

Generally finished creating ErrorPage in case of the ErrorBoundary being triggered.

Added NoMatch.js file to src/error. If a URL is entered that does not match a route it will trigger the ErrorBoundary using NoMatch component.

ErrorPage will not navigate if user tries to manually enter URL. Unsure how to resolve.