Slight style changes, incomplete bug fix.
Been having a bug for awhile that is being difficult to resolve. Specifically, on a mobile device the offcanvas sidebar won't navigate, unless you zoom out on it first. If navigating is attempted without it, it'll activate the clickHandler() under src/sidebar/SidebarList.js, but it won't navigate and focus will activate on an element behind the sidebar.  Works fine on desktop using the same browser.

Fixed the banner portion of src/layout/Header.js. Added a transparent <h1> element so that the banner doesn't cut off lower than the shown <h1> and the shown <h1> isn't effected by the mask.

The rest of the style changes are to improve look, while some of the other fixes such as the handleFocus() function are attempted fixes to earlier mentioned bug that improve user experience or are at least benign, such as moving the offcanvas to src/layout/Layout.js instead of keeping it in src/layout/MobileToolbar.js.