scrollIntoView() bug fixed, styling update, Description.js update
The bug that prevented scrollIntoView() from operating when using a mobile device was caused by having overflow: hidden for some of the components. scrollIntoView() is used to scroll to content using Offcanvas react-bootstrap component that rendered a sidebar. Restyled components to improve look and remove overflow, added overflow: auto to most if not all parent containers to prevent this from happening in the future. Removed Hit component and relevant variables/functions that was used to track events firing on a mobile device.

src/format/Description.js was giving warnings due to having <a> links within <a> links and due to not having a key prop on the top level component returned by map() function. Fixed both issues, although utilizing <object> around the nested <a> link is a arguably a bug since it's still nested. However it works and any other layout I could think of looks worse. Might be able do something with reference: https://stackoverflow.com/questions/9882916/are-you-allowed-to-nest-a-link-inside-of-a-link.