Website layout and styling improvements
When trying to style the projects section of Portfolio page, the card text made the layout of the web site look unprofessional. May go back in the future and possibly add the description back in, however enough time has been spent on this and it was easier to reformat as a button. Added scr/format/ViewButtonGroup.js and adjusted src/format/Item.js and List.js to accomplish this.

Change in how pdf files are rendered. embed element was being used to render pdf files, however getting them to style correctly proved to be a challenge. Instead embed was changed over to object element and rendered that way.

The rest of the changes were styling changes to improve the overall look of the site or to get the styling to format correctly on a cellular device as well as a PC.